initial commit, added helper/fixture folder and files
