+++ fixtures/testData.json
@@ -1,0 +1,8 @@
+{
+  "year": "2025",
+  "month": "08",
+  "day": "30",
+  "from": "New York",
+  "to": "Berlin",
+  "airport": "John F. Kennedy International AirportJFK"
+}
